Move graph construction into its own package
